1.指令扩展为int型  2.hero和monster的status，input分开
1. 指令扩展为32位，可满足32种指令类型
2.hero和monster的status根据实际逻辑单独写了一份，其中hero会多出来即时攻击和节拍卡点移动两张状态，只用hero有input， monster不再有input
3.界面上加了一个临时的节拍显示

+++ Assets/Scripts/Agent/Combo/ComboTrigger.cs
@@ -100,7 +100,7 @@ public class ComboTrigger : IMeterHandler
         }
     }
 
-    private TriggeredComboStep Trigger(byte newInput, int meterIndex)
+    private TriggeredComboStep Trigger(int newInput, int meterIndex)
     {
         TriggeredComboStep tcs = null;
 
@@ -137,7 +137,7 @@ public class ComboTrigger : IMeterHandler
     /// <param name="newInput">输入</param>
     /// <param name="meterIndex">输入对应的节拍index</param>
     /// <returns>触发combo的结果</returns>
-    public int OnNewInput(byte newInput, int meterIndex, out TriggeredComboStep triggeredComboStep)
+    public int OnNewInput(int newInput, int meterIndex, out TriggeredComboStep triggeredComboStep)
     {
         triggeredComboStep = null;
         if (meterIndex > comboLogicEndMeter && resetFlag)
